requirements.txt
about-time==4.2.1
alive-progress==3.1.1
anyio==3.6.2
attrs==23.1.0
autograd==1.5
botorch==0.8.3
click==8.1.3
cloudpickle==2.2.1
cma==3.2.2
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
deap==1.3.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
fastapi==0.95.1
fastprogress==0.2.6
fonttools==4.39.3
future==0.18.3
GPy==1.10.0
GPyOpt==1.2.6
gpytorch==1.9.1
grapheme==0.6.0
greenlet==2.0.2
h11==0.14.0
h5py==3.8.0
hyperopt==0.2.7
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
linear-operator==0.3.0
llvmlite==0.38.1
loguru==0.7.0
matplotlib==3.7.1
multipledispatch==0.6.0
networkx==3.1
numba==0.55.2
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
olymp==0.0.1b0
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
paramz==0.9.5
Pillow==9.5.0
pluggy==1.0.0
py==1.11.0
py4j==0.10.9.7
pydantic==1.10.7
pyDOE==0.3.8
pymoo==0.6.0.1
pyparsing==3.0.9
pyrff==2.0.2
pyro-api==0.1.2
pyro-ppl==1.8.4
pyswarms==1.3.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
skorch==0.11.0
sniffio==1.3.0
sobol-seq==0.2.0
SQCommon==0.3.2
SQLAlchemy==2.0.9
SQSnobFit==0.4.5
starlette==0.26.1
summit==0.8.9
tabulate==0.9.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
wrapt==1.15.0
silence_tensorflow
tensorflow
tensorflow_probability
